+++ aa_forum/views/admin.py
@@ -30,9 +30,9 @@ from app_utils.logging import LoggerAddTag
 # AA Forum
 from aa_forum import __title__
 from aa_forum.constants import DEFAULT_CATEGORY_AND_BOARD_SORT_ORDER
-from aa_forum.forms import EditBoardForm, EditCategoryForm, NewCategoryForm
+from aa_forum.forms import EditBoardForm, EditCategoryForm, NewCategoryForm, SettingForm
 from aa_forum.helper.forms import message_form_errors
-from aa_forum.models import Board, Category
+from aa_forum.models import Board, Category, Setting
 
 logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
@@ -498,6 +498,41 @@ def forum_settings(request: WSGIRequest) -> HttpResponse:
     :return:
     """
 
-    context = {}
+    logger.info(f"{request.user} called forum settings page")
+
+    settings = Setting.objects.get(pk=1)
+
+    # If this is a POST request we need to process the form data
+    if request.method == "POST":
+        settings_form = SettingForm(request.POST, instance=settings)
+
+        # Check whether it's valid:
+        if settings_form.is_valid():
+            settings.messages_per_page = settings_form.cleaned_data["messages_per_page"]
+            settings.topics_per_page = settings_form.cleaned_data["topics_per_page"]
+            settings.user_signature_length = settings_form.cleaned_data[
+                "user_signature_length"
+            ]
+            settings.save()
+
+            messages.success(
+                request, mark_safe(_("<h4>Success!</h4><p>Settings updated.<p>"))
+            )
+
+            return redirect("aa_forum:admin_forum_settings")
+        else:
+            messages.error(
+                request,
+                mark_safe(
+                    _(
+                        "<h4>Error!</h4>"
+                        "<p>Something went wrong, please check your input<p>"
+                    )
+                ),
+            )
+    else:
+        settings_form = SettingForm(instance=settings)
+
+    context = {"form": settings_form}
 
     return render(request, "aa_forum/view/administration/forum-settings.html", context)